Register columns on inline-save ajax action for quickedit
In #30, we changed the registration of the Publishing Checklist column
to happen on `admin_head`, so that we could have access to all
registered post types and checklist tasks. This breaks the rendering of
the checklist column in quickedit saves, because the ajax response for
the quickedit save doesn't run the `admin_head` hook.

This change brings the quickedit response back in line with the initial
list table rendering, by running the same function on
`wp_ajax_inline-save` as is run on `admin_head`.